rename: 패키지 변경사항을 test 패키지에 적용 (#426)
